stopping point
this all works except the second query fails. the for loop runs without waiting for the connection to return any info. shows inside the for loop counting up, then runs to the quit menu, THEN it runs the connection with the end number of the for loop 5 times and has undefined for data[i]